+++ kodando-react-dom/src/main/kotlin/kodando/react/dom/HTMLElementBuilder.kt
@@ -4,7 +4,7 @@ import kodando.react.*
 
 open class HTMLElementBuilder<out TProps : HTMLElementAttributes>(val tagName: String) : Builder<TProps> {
 
-    override fun build(setter: PropSetter<TProps>): ReactElement? {
+    override operator fun invoke(setter: PropSetter<TProps>): ReactElement? {
         return createElement(tagName, unsafePropsBy(setter))
     }