+++ kodando-react-router/src/main/kotlin/kodando/react/router/dom/BrowserRouter.kt
@@ -6,8 +6,8 @@ import kodando.react.*
  * Created by danfma on 24/04/17.
  */
 
-val browserRouterBuilder = ComponentBuilder(Module.BrowserRouterClass)
+val BrowserRouter = ComponentBuilder(Module.BrowserRouterClass)
 
 
 fun ReactProps.browserRouter(setter: PropSetter<BrowserRouterProps>) =
-    append(browserRouterBuilder.build(setter))
+    append(BrowserRouter(setter))