+++ kodando-react-router/src/main/kotlin/kodando/react/router/dom/Router.kt
@@ -6,9 +6,9 @@ import kodando.react.*
  * Created by danfma on 04/04/17.
  */
 
-val routerBuilder = ComponentBuilder(Module.RouterClass)
+val Router = ComponentBuilder(Module.RouterClass)
 
 
 fun ReactProps.router(setter: PropSetter<RouterProps>) {
-    append(routerBuilder.build(setter))
+    append(Router(setter))
 }