+++ kodando-react/src/main/kotlin/kodando/react/Ksx.kt
@@ -3,5 +3,5 @@ package kodando.react
 object Ksx : KsxBuilder()
 
 fun ksx(setter: PropSetter<ReactProps>): ReactElement? {
-    return Ksx.build(setter)
+    return Ksx(setter)
 }