+++ kodando-react/src/main/kotlin/kodando/react/KsxBuilder.kt
@@ -1,7 +1,7 @@
 package kodando.react
 
 open class KsxBuilder : Builder<ReactProps> {
-    override fun build(setter: PropSetter<ReactProps>): ReactElement? {
+    operator override fun invoke(setter: PropSetter<ReactProps>): ReactElement? {
         return unsafePropsBy(setter).children?.firstAsElement()
     }
 }