+++ samples/react-showroom/src/main/kotlin/showroom/Shell.kt
@@ -5,6 +5,7 @@ import kodando.react.dom.*
 import kodando.react.router.dom.*
 import showroom.time.clock
 
+
 object Shell : StatelessRenderer<Shell.Props>() {
 
     interface Props : ReactProps